move Report class into utils.reports
