+++ scripts/train_vggface2_cleverhans.py
@@ -32,14 +32,13 @@ from __future__ import unicode_literals
 import tensorflow as tf
 from tensorflow import keras
 import tensorflow.keras.backend as K
-import pandas as pd
 
 from cleverhans.compat import flags
 from cleverhans.utils_keras import KerasModelWrapper
 
 from dro.training.models import vggface2_model
-from dro.utils.training_utils import make_callbacks, make_model_uid, \
-    get_n_from_file_pattern, compute_n_train_n_val, \
+from dro.utils.reports import Report
+from dro.utils.training_utils import make_callbacks, get_n_from_file_pattern, compute_n_train_n_val, \
     steps_per_epoch
 from dro.datasets import ImageDataset
 from dro.utils.flags import define_training_flags, define_adv_training_flags
@@ -57,33 +56,6 @@ define_training_flags()
 define_adv_training_flags()
 
 flags.DEFINE_bool("train_mnist", False, "whether to train the cleverhans mnist model.")
-
-
-class Report:
-    def __init__(self):
-        self.results_list = list()
-        # Set is_adversarial=True when generating the model_uid so that the adversarial
-        # parameters (attack type, epsilon, etc) will be recorded in the uid.
-        self.uid = make_model_uid(FLAGS, is_adversarial=True)
-        self.metric = keys.ACC  # the name of the metric being recorded
-
-    def add_result(self, val, model, data, phase):
-        """Record the results of an experiment."""
-        results_entry = (self.uid, self.metric, val, model, data, phase)
-        # Check for duplicates; while this is not strictly a problem, it is almost
-        # definitely a mistake if duplicate results are being added.
-        assert results_entry not in self.results_list, "duplicate results added to report"
-        self.results_list.append(results_entry)
-        return
-
-    def to_csv(self):
-        df = pd.DataFrame(self.results_list, columns=["uid", "metric", "value",
-                                                      "model", "data", "phase"])
-        fp = "./metrics/{}.csv".format(self.uid)
-        print("[INFO] writing results to {}".format(fp))
-        print(df)
-        df.to_csv(fp, index=False)
-        return
 
 
 def mnist_tutorial(label_smoothing=0.1):
